# Conflicts:
#	app/Http/Controllers/Admin/AdminController.php
#	app/Http/Controllers/ReportController.php
#	app/Http/Controllers/RoleController.php
#	app/Http/Controllers/UserController.php

+++ app/Http/Controllers/Admin/AdminController.php
@@ -12,8 +12,10 @@ use App\Models\Opportunity;
 use App\Models\Post;
 use App\Models\Seeker;
 use App\Models\User;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 use Spatie\Activitylog\Models\Activity;
 
 class AdminController extends Controller
@@ -27,7 +29,6 @@ class AdminController extends Controller
                 return $this->apiResponse(null, __('strings.user_removed_successfully'), 200);
             }
             return $this->apiResponse(null, __('strings.not_found'), 404);
-
     }
 
     public function getUsers($type) {
@@ -48,11 +49,10 @@ class AdminController extends Controller
             $result = UserResource::collection($seekers);
         }
 
-        else if($type == 'Companies' ) {
-            $seekers = User::role('company')->latest()->get();
-            $result = UserResource::collection($seekers);
+        else if($type == 'Companies') {
+            $companies = User::role('company')->latest()->get();
+            $result = UserResource::collection($companies);
         }
-
         else {
             return $this->apiResponse(null, __('strings.error_user_type'), 403);
         }
@@ -90,12 +90,28 @@ class AdminController extends Controller
     }
 
     public function banUser(Request $request, $id){
+        $vaildate = Validator::make($request->all(), [
+            'reason' => 'required',
+            'type' => 'required'
+        ]);
+        if ($vaildate->fails()) {
+            return $this->apiResponse(null, $vaildate->errors(), 400);
+        }
         $user = User::find($id);
         if ($user->isNotBanned()) {
+            $comment = $request->comment;
+            $type = $request->type;
+            $expired_at = $request->expired_at;
+            if ($type === 'forever') {
+                $expired_at = null;
+            } else if (!$expired_at) {
+                return $this->apiResponse(null, 'Date is required', 400);
+            }
+
             $user->roles()->detach();
             $ban = $user->ban([
-                'comment' => $request->comment,
-                'expired_at' => $request->expired_at
+                'comment' => $comment,
+                'expired_at' => $expired_at
             ]);
             return $this->apiResponse($ban, __('strings.banned_successfully'), 200);
         } else {
@@ -108,7 +124,6 @@ class AdminController extends Controller
         if ($user->isBanned()) {
             $user->syncRoles($user->roles_name);
             $user->unBan();
-
             return $this->apiResponse(null, __('strings.unbanned_successfully'), 200);
         } else {
             return $this->apiResponse(null, __('strings.user_already_not_banned'), 403);
@@ -129,19 +144,20 @@ class AdminController extends Controller
             }
         });
         return $this->apiResponse($users, __('strings.all_users_banned'), 200);
+        return $this->apiResponse(UserResource::collection($users), __('strings.all_users_banned'), 200);
     }
 
-    public function deleteExpiredBanned() {
-        $users = User::onlyBanned()->get();
-        $users = $users->reject(function(User $user) {
-            $roles = $user->roles_name;
-            foreach ($roles as $value) {
-                return $value === 'owner';
-            }
-        });
-
-        return $this->apiResponse($users, __('strings.all_users_banned'), 200);
-    }
+    // public function deleteExpiredBanned() {
+    //     $users = User::onlyBanned()->get();
+    //     $user = User::find(6);
+    //     return $user->bans;
+    //     foreach ($users as $user) {
+    //         if ($user->bans[0]->expired_at < Carbon::now()) {
+    //             $user->syncRoles($user->roles_name);
+    //             $user->unBan();
+    //         }
+    //     }
+    // }
 
     public function countPOA() {
         $posts = Post::count();
@@ -175,13 +191,33 @@ class AdminController extends Controller
     }
 
     public function lineChart() {
-        $logs = Activity::all();
-        $data = [];
-        $count = 0;
-        foreach ($logs as $value) {
+        $minDate = Activity::min('created_at');
+        $maxDate = Activity::max('created_at');
 
-            $data[$value->created_at->format('M')] = $value->created_at->format('D-M-Y');
+        if (!$minDate || !$maxDate) {
+            return response()->json(['message' => 'No data available'], 404);
         }
-        return $data;
+
+        $startDate = Carbon::parse($minDate)->startOfDay();
+        $endDate = Carbon::parse($maxDate)->endOfDay();
+
+        $dailyData = [];
+
+        while ($startDate->lte($endDate)) {
+            $dayStart = $startDate->copy()->startOfDay()->toDateTimeString();
+            $dayEnd = $startDate->copy()->endOfDay()->toDateTimeString();
+
+            $count = Activity::whereBetween('created_at', [$dayStart, $dayEnd])->count();
+            $dayFormat = $startDate->copy()->startOfDay()->format('M-d');
+            $dailyData[] = [
+                'day' => $dayFormat,
+                'count' => $count,
+                'amt' => 2000
+            ];
+
+            $startDate->addDay();
+        }
+
+        return response()->json($dailyData);
     }
 }
